% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_truncated_multivariate_normal.R
\name{Gibbs_truncated_multivariate_normal}
\alias{Gibbs_truncated_multivariate_normal}
\title{Gibbs sampler for a truncated multivariate normal distribution}
\usage{
Gibbs_truncated_multivariate_normal(
  n = 1,
  alpha,
  beta,
  gamma,
  Gamma,
  which = "Z",
  steps = 20
)
}
\arguments{
\item{n}{Number of samples. Default: \code{n=1}.}

\item{alpha}{Vector of lower bounds for S component.}

\item{beta}{Vector of upper bounds for S component.}

\item{gamma}{Matrix of linear transformations on Z component.}

\item{Gamma}{Cholesky factor for variance on Z component.}

\item{which}{Text string (\code{"S"/"Z"/"SZ"}) deciding which componts are return. Default: \code{which="Z"}.}

\item{steps}{Number of steps in Gibbs sampler. Default: \code{steps=20}.}
}
\value{
Simulations either as a matrix (if \code{which="S"} or \code{"Z"}), or as a list of matrices (if \code{which="SZ"}).
}
\description{
Simulate \code{S in R^M} and \code{Z in R^d} given \code{S in (alpha,beta]$}, where \code{S = X + gamma^T Z} with \code{X ~ Normal(0,I_N)} and \code{Z ~ Normal(0,Gamma^T Gamma)}.
}
\examples{
Gibbs_truncated_multivariate_normal(n=10,alpha=rep(-1,5),beta=rep(1,5),gamma=matrix(rnorm(10),2,5),Gamma=matrix(c(1,1,0,1),2,2))
}
