% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_slim.R
\name{predict_slim}
\alias{predict_slim}
\title{Prediction from new dataset}
\usage{
predict_slim(object, newdata)
}
\arguments{
\item{object}{of clasee \code{\link[stats]{lm}} or \code{\link[biglm]{biglm}}.}

\item{newdata}{data frame on which predictions should be made.}
}
\value{
vector with predictions.
}
\description{
Predictions from a linear model.
}
\details{
A slim implementation of predictions from a linear model
on a new dataset. This is made in replacement of \code{\link[stats]{predict}}
and \code{\link[biglm]{predict.bigglm}}, which do not work on
\code{\link[biglm]{biglm}}-objects or does not include a possible
\code{\link[stats]{offset}}, respectively.
}
\note{
The implementation uses \code{biglm:::coef.biglm} as this function
is not exported from the \code{\link[biglm]{biglm}}-package.
}
