% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_probit.R
\name{probit-class}
\alias{probit-class}
\alias{print.probit}
\alias{summary.probit}
\alias{anova.probit}
\title{The \code{probit} class and some basic methods}
\usage{
\method{print}{probit}(x)

\method{summary}{probit}(x)

\method{anova}{probit}(x, y)
}
\arguments{
\item{x}{Object of class \code{probit}.}

\item{y}{Object of class \code{probit}.}
}
\description{
Objects of class \code{probit} as generated by \code{\link{probit}} is a list with entries as specified below.
\describe{
  \item{\code{regression}}{List indexed by \code{items} of objects of class \code{clm} fitted by \code{\link[ordinal]{clm}}}
  \item{\code{Gamma}}{Cholesky factor of estimated variance of random effects}
  \item{\code{dependence}}{Text string (\code{marginal} or \code{joint}) specifying the used correlation structure}
  \item{\code{items}}{Vector with item names}
  \item{\code{subject}}{Name of subject variable inside data}
  \item{\code{random.eff}}{Vector with names of random effects}
  \item{\code{data}}{Tibble with dataset analyzed}
}
}
