% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit}
\alias{probit}
\title{Probit analysis with random effects}
\usage{
probit(
  formula,
  random,
  subject = "id",
  dependence = "marginal",
  Gamma = NULL,
  item.name = "item",
  response.name = NULL,
  data,
  data.long = NULL,
  B = 300,
  BB = 50,
  maxit = 20,
  sig.level = 0.6,
  verbose = 0
)
}
\arguments{
\item{formula}{Model formula, where multivariate responses may be given additively on the left hand side. Responses must be ordered factors.}

\item{random}{List of formulas for the random effects. Models are fitted on first appearance observations.}

\item{subject}{Categorical variable encoding subjects.}

\item{dependence}{Text string (\code{"marginal" or "joint"}) deciding whether random effects are assumed independent or with a common joint normal distribution. Default: \code{dependence="marginal"}.}

\item{Gamma}{Choleskey factor for initial variance of random effects. If \code{Gamma=NULL} then initialized at identity matrix. Default: \code{Gamma=NULL}.}

\item{item.name}{Character string with name of generated variable identifying the items. Defaults to \code{item.name="item"}.}

\item{response.name}{Character string with name of generated variable containing the responses. Defaults to \code{response.name=NULL}, which corresponds to adding \code{".value"} to \code{item.name}.}

\item{data}{Date frame with data on the wide format.}

\item{data.long}{Possible additional long format data frame with item level explanatory variables. Presently not implemented!}

\item{B}{Number of simulations in minimization step. Default: \code{B=300}.}

\item{BB}{Number of simulations per subject in maximization step. Default: \code{BB=50}.}

\item{maxit}{Maximal number of minimization-maximization steps. Default: \code{steps=20}.}

\item{sig.level}{Significance level at which the iterative stochastic optimizations will be stopped. Defaults to \code{sig.level=0.60}.}

\item{verbose}{Numeric controlling amount of convergence diagnostics. Default: \code{verbose=0} corresponding to no output.}
}
\value{
Object of class \code{probit}.
}
\description{
Separate analysis over items.
}
\details{
If \code{Gamma=NULL} then predicted random effects are initialized at zero. Otherwise they are simulated from normal distribution with variance \code{Gamma^T Gamma}.
}
\note{
A data frame must be provided, i.e. the \code{data} option is not optional. Variables not appearing in the \code{data}, but appearing in the \code{formula}, will be assumed to be random.
}
