% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_old.R
\name{probit_old}
\alias{probit_old}
\title{Probit analysis with random effects}
\usage{
probit_old(
  formula,
  subject,
  data,
  dependence = "marginal",
  Gamma = NULL,
  M = 10,
  EMsteps = 20,
  eps = 1e-04,
  maxit = 50,
  verbose = FALSE
)
}
\arguments{
\item{formula}{Model formula, where multivariate responses may be given additively on the left hand side. Responses must be ordered factors.}

\item{subject}{Categorical variable encoding subjects.}

\item{data}{Date frame with data.}

\item{dependence}{Text string (\code{"marginal" or "joint"}) deciding whether random effects are assumed independent or with a common joint normal distribution. Default: \code{dependence="marginal"}.}

\item{Gamma}{Choleskey factor for initial variance of random effects. If \code{Gamma=NULL} then initialized at identity matrix. Default: \code{Gamma=NULL}.}

\item{M}{Number of replications in simulation of E-step. Default: \code{M=10}.}

\item{EMsteps}{Maximal number of EM steps. Default: \code{EMsteps=20}.}

\item{eps}{Absolute convergence criterion for total log likelihood. Default: \code{eps=1e-4}.}

\item{maxit}{Maximal number of steps in underlying coupling from the past algorithm. Default: \code{maxit=50}.}

\item{verbose}{If TRUE, then print convergence diagnostics. Default: \code{verbose=FALSE}.}
}
\value{
Object of class \code{probit}.
}
\description{
Separate analysis over items.
}
\details{
If \code{Gamma=NULL} then predicted random effects are initialized at zero. Otherwise they are simulated from normal distribution with variance \code{Gamma^T Gamma}.
}
\note{
A data frame must be provided, i.e. the \code{data} option is not optional. Variables not appearing in the \code{data}, but appearing in the \code{formula}, will be assumed to be random.
}
