% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef_probit.R
\name{ranef.probit}
\alias{ranef.probit}
\title{Simulate random effects}
\usage{
\method{ranef}{probit}(x, maxit = 500)
}
\arguments{
\item{x}{Object of class \code{\link{probit-class}}}

\item{maxit}{Maximal number of steps in coupling from the past algorithm. Default: \code{maxit=500}.}
}
\description{
Simulate random effects from a probit object.
}
