% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_probit.R
\name{simulate.probit}
\alias{simulate.probit}
\title{update method for probit-class}
\usage{
\method{simulate}{probit}(
  object,
  nsim = 1,
  seed = NULL,
  re.form = NULL,
  re.keep = TRUE,
  make.ordinal = TRUE,
  data = NULL
)
}
\arguments{
\item{object}{Object of class \code{probit}.}

\item{nsim}{Number of response vectors to simulate. Defaults to \code{1}.}

\item{seed}{An object specifying if and how the random number generator should be initialized (‘seeded’). Defaults to \code{NULL}. See \code{\link{simulate}} for more information.}

\item{re.form}{(\code{formula}, \code{NULL}, or \code{NA}) specify which random effects to include in simulation. Defaults to \code{NULL} in which case all random effects are included. If \code{NA} or \code{~0} no random effects are included.}

\item{re.keep}{Keep realizations of random effects in the model? Defaults to \code{TRUE}. If \code{FALSE} and also \code{nsim > 1}, then simulations will be stored as lists of vectors.}

\item{make.ordinal}{Should ordinal responses be categorized from their numerical value? Defaults to \code{TRUE}.}

\item{data}{Date frame with data on the wide format. Defaults to \code{data=NULL}, which corresponds to \code{data} taken from the call object.}
}
\value{
\code{\link{probit-class}} object.
}
\description{
simulate method for \code{probit}-object as generated by \code{\link{probit}}.
}
