% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_deep.R
\name{update_deep}
\alias{update_deep}
\title{update method for probit-class}
\usage{
update_deep(
  object,
  fixed = NULL,
  random = NULL,
  A = NULL,
  dependence = NULL,
  data = NULL,
  B = NULL,
  BB = NULL,
  maxit = 20,
  sig.level = 0.6,
  verbose = 0,
  estimate.models = TRUE
)
}
\arguments{
\item{object}{Object of class \code{probit}.}

\item{fixed}{Update of right hand side of model formula for fixed effects. Defaults to \code{fixed=NULL}, which corresponds to no update.}

\item{random}{List of updates of right hand side of formulas for the random effects. Defaults to \code{random=NULL}, which corresponds to no update.}

\item{A}{Transformation matrix of predicted random effects. Defaults to \code{A=NULL}, which corresponds to identity transformation of residing random effects.}

\item{dependence}{Text string (\code{"marginal" or "joint"}) deciding whether random effects are assumed independent or with a common joint normal distribution. Defaults to \code{dependence=NULL}, which corresponds to no update.}

\item{data}{Date frame with data on the wide format. Defaults to \code{data=NULL}, which corresponds to \code{data} taken from the call object.}

\item{B}{Number of simulations in minimization step. Defaults to \code{B=NULL}, which corresponds to \code{B} taken from the call object.}

\item{BB}{Number of simulations per subject in \code{anova} or \code{update}. Defaults to \code{BB=NULL}, which corresponds to \code{BB} taken from the call object.}

\item{maxit}{Maximal number of minimization-maximization steps. Defaults to \code{maxit=20}.}

\item{sig.level}{Significance level at which the iterative stochastic optimization will be stopped. Defaults to \code{sig.level=0.60}.}

\item{verbose}{Numeric controlling amount of convergence diagnostics. Default: \code{verbose=0} corresponding to no output.}

\item{estimate.models}{Boolean deciding if model parameters are estimated in \code{update}.}
}
\value{
\code{\link{probit-class}} object.
}
\description{
update method for \code{probit}-object as generated by \code{\link{probit}}.
}
