% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_probit.R
\name{probit-class}
\alias{probit-class}
\alias{print.probit}
\alias{summary.probit}
\alias{confint.probit}
\alias{anova.probit}
\alias{print.probit_anova}
\alias{update.probit}
\alias{recover_data.probit}
\alias{emm_basis.probit}
\title{The \code{probit} class and some basic methods}
\usage{
\method{print}{probit}(x)

\method{summary}{probit}(object)

\method{confint}{probit}(object, level = 0.95)

\method{anova}{probit}(object, BB = NULL)

\method{print}{probit_anova}(x, digits = 4)

\method{update}{probit}(
  object,
  fixed = NULL,
  random = NULL,
  dependence = NULL,
  data = NULL,
  B = NULL,
  BB = NULL,
  maxit = 20,
  sig.level = 0.6,
  verbose = 0
)

\method{recover_data}{probit}(object, ...)

\method{emm_basis}{probit}(object, trms, xlev, grid, ...)
}
\arguments{
\item{x}{Object of class \code{probit}.}

\item{object}{Object of class \code{probit}.}

\item{level}{Approximative coverage for confidence interval. Defaults to \code{level=0.95}.}

\item{BB}{Number of simulations per subject in \code{anova} or \code{update}. Defaults to \code{BB=NULL}, which corresponds to \code{BB} taken from the call object.}

\item{digits}{Number of digits in print of \code{probit_anova} object. Defaults to \code{digits=4}.}

\item{fixed}{Update of right hand side of model formula for fixed effects. Defaults to \code{fixed=NULL}, which corresponds to no update.}

\item{random}{List of updates of right hand side of formulas for the random effects. Defaults to \code{random=NULL}, which corresponds to no update.}

\item{dependence}{Text string (\code{"marginal" or "joint"}) deciding whether random effects are assumed independent or with a common joint normal distribution. Defaults to \code{dependence=NULL}, which corresponds to no update.}

\item{data}{Date frame with data on the wide format. Defaults to \code{data=NULL}, which corresponds to \code{data} taken from the call object.}

\item{B}{Number of simulations in minimization step. Defaults to \code{B=NULL}, which corresponds to \code{B} taken from the call object.}

\item{maxit}{Maximal number of minimization-maximization steps. Defaults to \code{maxit=20}.}

\item{sig.level}{Significance level at which the iterative stochastic optimizations will be stopped. Defaults to \code{sig.level=0.60}.}

\item{verbose}{Numeric controlling amount of convergence diagnostics. Default: \code{verbose=0} corresponding to no output.}
}
\description{
Objects of class \code{probit} as generated by \code{\link{probit}} is a list with entries as specified below.
\describe{
  \item{\code{call}}{Function call.}
  \item{\code{response.name}}{Name of response in internal representation.}
  \item{\code{item.name}}{Name of item identifier in the model formula.}
  \item{\code{items.interval}}{Character vector with names of interval responses.}
  \item{\code{items.ordinal}}{Character vector with names of ordinal responses.}
  \item{\code{subject}}{Name of subject identifier.}
  \item{\code{random}}{List of mean models for the random effects.}
  \item{\code{dependence}}{Text string (\code{marginal} or \code{joint}) specifying the used correlation structure.}
  \item{\code{m.fixed}}{\code{\link[biglm]{biglm}}-object with joint fit of fixed effects. Note that test and confidence intervals on this object does not make direct sense.}
  \item{\code{sigma2}}{List with estimated variances for the interval responses.}
  \item{\code{eta}}{List with estimated threshold parameters for the ordinal responses.}
  \item{\code{m.random}}{List of \code{\link[stats]{lm}}-objects for mean models on the random effects.}
  \item{\code{Gamma}}{Cholesky factor of estimated inverse variance of random effects.}
  \item{\code{mu}}{Matrix (subjects, q) of estimated approximate conditional means for the random effects, where q = number of random effects.}
  \item{\code{psi}}{Matrix (subjects, q*(q+1)/2) of estimated Cholesky factors of approximate inverse conditional variances for the random effects.}
  \item{\code{B}}{Number of replications for each subject in Monte-Carlo computation in minimization step.}
  \item{\code{BB}}{Number of replications for each subject in Monte-Carlo computation in maximization step.}
  \item{\code{F1}}{Vector with subject-wise \code{F1}-values from last minimization step.}
  \item{\code{pvalue}}{p-value from latest T-test comparing the stochastic optimization.}
  \item{\code{iter}}{Number of minimization-maximization steps.}
  \item{\code{code}}{Iteration status: 0=iterations halted due to p-value, 1=iterations reached maximum iterations.}
  \item{\code{data}}{Tibble with dataset analyzed.}
}
}
\note{
\code{anova} re-simulates the underlying responses and random
effects for the fixed effects model. Hence the output of the top part
of anova table is stochastic.
}
