% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit}
\alias{probit}
\title{Probit analysis with random effects}
\usage{
probit(formula, item, subject, data, dependence = "marginal")
}
\arguments{
\item{formula}{Model formula to be used separately for each item.}

\item{item}{Categorical variable encoding items.}

\item{subject}{Categorical variable encoding subjects.}

\item{data}{Date frame with data.}

\item{dependence}{Text string (\code{"marginal"/"joint"}) deciding whether random effects are assumed independent or with a common joint normal distribution. Default: \code{dependence="marginal"}.}
}
\value{
Object of class \code{probit}.
}
\description{
Separate analysis over item.
}
\note{
A data frame must be provided, i.e. the \code{data} option is not optional. Variables not appearing in the \code{data}, but appearing in the \code{formula}, will be assumed to be random.
}
